use-remote-details tests
